// src/app/spaceTraderAPI/api/api.ts
export interface Cooldown {
    /**
     * The symbol of the ship that is on cooldown
     * @type {string}
     * @memberof Cooldown
     */
    'shipSymbol': string;
    /**
     * The total duration of the cooldown in seconds
     * @type {number}
     * @memberof Cooldown
     */
    'totalSeconds': number;
    /**
     * The remaining duration of the cooldown in seconds
     * @type {number}
     * @memberof Cooldown
     */
    'remainingSeconds': number;
    /**
     * The date and time when the cooldown expires in ISO 8601 format
     * @type {string}
     * @memberof Cooldown
     */
    'expiration'?: string;
}
// ...
export interface Ship {
    /**
     * The globally unique identifier of the ship in the following format: `[AGENT_SYMBOL]-[HEX_ID]`
     * @type {string}
     * @memberof Ship
     */
    'symbol': string;
    /**
     * 
     * @type {ShipRegistration}
     * @memberof Ship
     */
    'registration': ShipRegistration;
    /**
     * 
     * @type {ShipNav}
     * @memberof Ship
     */
    'nav': ShipNav;
    /**
     * 
     * @type {ShipCrew}
     * @memberof Ship
     */
    'crew': ShipCrew;
    /**
     * 
     * @type {ShipFrame}
     * @memberof Ship
     */
    'frame': ShipFrame;
    /**
     * 
     * @type {ShipReactor}
     * @memberof Ship
     */
    'reactor': ShipReactor;
    /**
     * 
     * @type {ShipEngine}
     * @memberof Ship
     */
    'engine': ShipEngine;
    /**
     * Modules installed in this ship.
     * @type {Array<ShipModule>}
     * @memberof Ship
     */
    'modules': Array<ShipModule>;
    /**
     * Mounts installed in this ship.
     * @type {Array<ShipMount>}
     * @memberof Ship
     */
    'mounts': Array<ShipMount>;
    /**
     * 
     * @type {ShipCargo}
     * @memberof Ship
     */
    'cargo': ShipCargo;
    /**
     * 
     * @type {ShipFuel}
     * @memberof Ship
     */
    'fuel': ShipFuel;
    /**
     * 
     * @type {Cooldown}
     * @memberof Ship
     */
    'cooldown': Cooldown;
}
/**
 * Ship cargo details.
 * @export
 * @interface ShipCargo
 */
export interface ShipCargo {
    /**
     * The max number of items that can be stored in the cargo hold.
     * @type {number}
     * @memberof ShipCargo
     */
    'capacity': number;
    /**
     * The number of items currently stored in the cargo hold.
     * @type {number}
     * @memberof ShipCargo
     */
    'units': number;
    /**
     * The items currently in the cargo hold.
     * @type {Array<ShipCargoItem>}
     * @memberof ShipCargo
     */
    'inventory': Array<ShipCargoItem>;
}
// ...
export interface ShipCrew {
    /**
     * The current number of crew members on the ship.
     * @type {number}
     * @memberof ShipCrew
     */
    'current': number;
    /**
     * The minimum number of crew members required to maintain the ship.
     * @type {number}
     * @memberof ShipCrew
     */
    'required': number;
    /**
     * The maximum number of crew members the ship can support.
     * @type {number}
     * @memberof ShipCrew
     */
    'capacity': number;
    /**
     * The rotation of crew shifts. A stricter shift improves the ship\'s performance. A more relaxed shift improves the crew\'s morale.
     * @type {string}
     * @memberof ShipCrew
     */
    'rotation': ShipCrewRotationEnum;
    /**
     * A rough measure of the crew\'s morale. A higher morale means the crew is happier and more productive. A lower morale means the ship is more prone to accidents.
     * @type {number}
     * @memberof ShipCrew
     */
    'morale': number;
    /**
     * The amount of credits per crew member paid per hour. Wages are paid when a ship docks at a civilized waypoint.
     * @type {number}
     * @memberof ShipCrew
     */
    'wages': number;
}

export const ShipCrewRotationEnum = {
    Strict: 'STRICT',
    Relaxed: 'RELAXED'
} as const;

export type ShipCrewRotationEnum = typeof ShipCrewRotationEnum[keyof typeof ShipCrewRotationEnum];

/**
 * The engine determines how quickly a ship travels between waypoints.
 * @export
 * @interface ShipEngine
 */
export interface ShipEngine {
    /**
     * The symbol of the engine.
     * @type {string}
     * @memberof ShipEngine
     */
    'symbol': ShipEngineSymbolEnum;
    /**
     * The name of the engine.
     * @type {string}
     * @memberof ShipEngine
     */
    'name': string;
    /**
     * The repairable condition of a component. A value of 0 indicates the component needs significant repairs, while a value of 1 indicates the component is in near perfect condition. As the condition of a component is repaired, the overall integrity of the component decreases.
     * @type {number}
     * @memberof ShipEngine
     */
    'condition': number;
    /**
     * The overall integrity of the component, which determines the performance of the component. A value of 0 indicates that the component is almost completely degraded, while a value of 1 indicates that the component is in near perfect condition. The integrity of the component is non-repairable, and represents permanent wear over time.
     * @type {number}
     * @memberof ShipEngine
     */
    'integrity': number;
    /**
     * The description of the engine.
     * @type {string}
     * @memberof ShipEngine
     */
    'description': string;
    /**
     * The speed stat of this engine. The higher the speed, the faster a ship can travel from one point to another. Reduces the time of arrival when navigating the ship.
     * @type {number}
     * @memberof ShipEngine
     */
    'speed': number;
    /**
     * 
     * @type {ShipRequirements}
     * @memberof ShipEngine
     */
    'requirements': ShipRequirements;
    /**
     * The overall quality of the component, which determines the quality of the component. High quality components return more ships parts and ship plating when a ship is scrapped. But also require more of these parts to repair. This is transparent to the player, as the parts are bought from/sold to the marketplace.
     * @type {number}
     * @memberof ShipEngine
     */
    'quality': number;
}
// ...
export interface ShipFrame {
    /**
     * Symbol of the frame.
     * @type {string}
     * @memberof ShipFrame
     */
    'symbol': ShipFrameSymbolEnum;
    /**
     * Name of the frame.
     * @type {string}
     * @memberof ShipFrame
     */
    'name': string;
    /**
     * The repairable condition of a component. A value of 0 indicates the component needs significant repairs, while a value of 1 indicates the component is in near perfect condition. As the condition of a component is repaired, the overall integrity of the component decreases.
     * @type {number}
     * @memberof ShipFrame
     */
    'condition': number;
    /**
     * The overall integrity of the component, which determines the performance of the component. A value of 0 indicates that the component is almost completely degraded, while a value of 1 indicates that the component is in near perfect condition. The integrity of the component is non-repairable, and represents permanent wear over time.
     * @type {number}
     * @memberof ShipFrame
     */
    'integrity': number;
    /**
     * Description of the frame.
     * @type {string}
     * @memberof ShipFrame
     */
    'description': string;
    /**
     * The amount of slots that can be dedicated to modules installed in the ship. Each installed module take up a number of slots, and once there are no more slots, no new modules can be installed.
     * @type {number}
     * @memberof ShipFrame
     */
    'moduleSlots': number;
    /**
     * The amount of slots that can be dedicated to mounts installed in the ship. Each installed mount takes up a number of points, and once there are no more points remaining, no new mounts can be installed.
     * @type {number}
     * @memberof ShipFrame
     */
    'mountingPoints': number;
    /**
     * The maximum amount of fuel that can be stored in this ship. When refueling, the ship will be refueled to this amount.
     * @type {number}
     * @memberof ShipFrame
     */
    'fuelCapacity': number;
    /**
     * 
     * @type {ShipRequirements}
     * @memberof ShipFrame
     */
    'requirements': ShipRequirements;
    /**
     * The overall quality of the component, which determines the quality of the component. High quality components return more ships parts and ship plating when a ship is scrapped. But also require more of these parts to repair. This is transparent to the player, as the parts are bought from/sold to the marketplace.
     * @type {number}
     * @memberof ShipFrame
     */
    'quality': number;
}
// ...
export interface ShipFuel {
    /**
     * The current amount of fuel in the ship\'s tanks.
     * @type {number}
     * @memberof ShipFuel
     */
    'current': number;
    /**
     * The maximum amount of fuel the ship\'s tanks can hold.
     * @type {number}
     * @memberof ShipFuel
     */
    'capacity': number;
    /**
     * 
     * @type {ShipFuelConsumed}
     * @memberof ShipFuel
     */
    'consumed'?: ShipFuelConsumed;
}
// ...
export interface ShipNav {
    /**
     * The symbol of the system.
     * @type {string}
     * @memberof ShipNav
     */
    'systemSymbol': string;
    /**
     * The symbol of the waypoint.
     * @type {string}
     * @memberof ShipNav
     */
    'waypointSymbol': string;
    /**
     * 
     * @type {ShipNavRoute}
     * @memberof ShipNav
     */
    'route': ShipNavRoute;
    /**
     * 
     * @type {ShipNavStatus}
     * @memberof ShipNav
     */
    'status': ShipNavStatus;
    /**
     * 
     * @type {ShipNavFlightMode}
     * @memberof ShipNav
     */
    'flightMode': ShipNavFlightMode;
}
// ...
export const ShipNavStatus = {
    InTransit: 'IN_TRANSIT',
    InOrbit: 'IN_ORBIT',
    Docked: 'DOCKED'
} as const;

export type ShipNavStatus = typeof ShipNavStatus[keyof typeof ShipNavStatus];


/**
 * The reactor of the ship. The reactor is responsible for powering the ship\'s systems and weapons.
 * @export
 * @interface ShipReactor
 */
export interface ShipReactor {
    /**
     * Symbol of the reactor.
     * @type {string}
     * @memberof ShipReactor
     */
    'symbol': ShipReactorSymbolEnum;
    /**
     * Name of the reactor.
     * @type {string}
     * @memberof ShipReactor
     */
    'name': string;
    /**
     * The repairable condition of a component. A value of 0 indicates the component needs significant repairs, while a value of 1 indicates the component is in near perfect condition. As the condition of a component is repaired, the overall integrity of the component decreases.
     * @type {number}
     * @memberof ShipReactor
     */
    'condition': number;
    /**
     * The overall integrity of the component, which determines the performance of the component. A value of 0 indicates that the component is almost completely degraded, while a value of 1 indicates that the component is in near perfect condition. The integrity of the component is non-repairable, and represents permanent wear over time.
     * @type {number}
     * @memberof ShipReactor
     */
    'integrity': number;
    /**
     * Description of the reactor.
     * @type {string}
     * @memberof ShipReactor
     */
    'description': string;
    /**
     * The amount of power provided by this reactor. The more power a reactor provides to the ship, the lower the cooldown it gets when using a module or mount that taxes the ship\'s power.
     * @type {number}
     * @memberof ShipReactor
     */
    'powerOutput': number;
    /**
     * 
     * @type {ShipRequirements}
     * @memberof ShipReactor
     */
    'requirements': ShipRequirements;
    /**
     * The overall quality of the component, which determines the quality of the component. High quality components return more ships parts and ship plating when a ship is scrapped. But also require more of these parts to repair. This is transparent to the player, as the parts are bought from/sold to the marketplace.
     * @type {number}
     * @memberof ShipReactor
     */
    'quality': number;
}
// ...
export interface ShipRegistration {
    /**
     * The agent\'s registered name of the ship
     * @type {string}
     * @memberof ShipRegistration
     */
    'name': string;
    /**
     * The symbol of the faction the ship is registered with
     * @type {string}
     * @memberof ShipRegistration
     */
    'factionSymbol': string;
    /**
     * 
     * @type {ShipRole}
     * @memberof ShipRegistration
     */
    'role': ShipRole;
}
// ...
export type ShipRole = typeof ShipRole[keyof typeof ShipRole];